organisation/serializer.py
from rest_framework import serializers
from organisation.models import Organisation

class OrganisationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Organisation
        fields = "__all__"
